requirements.txt
anthropic>=0.39.0
mcp>=1.0.0
streamlit>=1.28.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
python-dotenv>=1.0.0
